Конвертує гривню в долар

+++ homework.py
@@ -1,4 +1,4 @@
-UAH = int(input('Введіть кількість гривень: '))
+UAH = float(input('Введіть кількість гривень: '))
 DOLLAR = 36.65 #ціна одного долара
 convert = round(UAH / DOLLAR, 2)
 print(convert, '$')
